C++ Program
Program user enters the age and based on the input, the if..else statement checks whether the entered age is greater than or equal to 18. If this condition meet then displays message "You are eligible for voting", however if the condition doesn't meet then displays a different message "You are not eligible for voting".